Move back `stateIcon`
