+++ source/features/list-pr-for-branch.tsx
@@ -5,8 +5,7 @@ import * as pageDetect from 'github-url-detection';
 import features from '.';
 import getDefaultBranch from '../github-helpers/get-default-branch';
 import addAfterBranchSelector from '../helpers/add-after-branch-selector';
-import {getPullRequestsAssociatedWithBranch} from './show-associated-branch-prs-on-fork';
-import {stateIcon} from '../github-helpers';
+import {getPullRequestsAssociatedWithBranch, stateIcon} from './show-associated-branch-prs-on-fork';
 
 // Taken from https://github.com/fregante/github-issue-link-status/blob/98792f2837352bacbf80664f3edbcec8e579ed17/source/github-issue-link-status.js#L10
 const stateColorMap = {
@@ -49,7 +48,8 @@ async function init(): Promise<void | false> {
 			data-hovercard-type="pull_request"
 			data-hovercard-url={prInfo.url + '/hovercard'}
 		>
-			<StateIcon className={stateColorMap[prInfo.state]}/> #{prInfo.number}
+			<StateIcon className={stateColorMap[prInfo.state]}/>
+			<span> #{prInfo.number}</span>
 		</a>
 	);
