+++ source/features/show-associated-branch-prs-on-fork.tsx
@@ -3,10 +3,11 @@ import cache from 'webext-storage-cache';
 import onetime from 'onetime';
 import {observe} from 'selector-observer';
 import * as pageDetect from 'github-url-detection';
+import {GitMergeIcon, GitPullRequestIcon, GitPullRequestClosedIcon, GitPullRequestDraftIcon} from '@primer/octicons-react';
 
 import features from '.';
 import * as api from '../github-helpers/api';
-import {getRepo, upperCaseFirst, stateIcon} from '../github-helpers';
+import {getRepo, upperCaseFirst} from '../github-helpers';
 
 interface PullRequest {
 	timelineItems: {
@@ -60,6 +61,13 @@ export const getPullRequestsAssociatedWithBranch = cache.function(async (): Prom
 	staleWhileRevalidate: {days: 4},
 	cacheKey: () => 'associatedBranchPullRequests:' + getRepo()!.nameWithOwner,
 });
+
+export const stateIcon = {
+	OPEN: GitPullRequestIcon,
+	CLOSED: GitPullRequestClosedIcon,
+	MERGED: GitMergeIcon,
+	DRAFT: GitPullRequestDraftIcon,
+};
 
 async function init(): Promise<void> {
 	const associatedPullRequests = await getPullRequestsAssociatedWithBranch();
