+++ source/github-helpers/index.ts
@@ -3,7 +3,6 @@ import onetime from 'onetime';
 import elementReady from 'element-ready';
 import compareVersions from 'tiny-version-compare';
 import * as pageDetect from 'github-url-detection';
-import {GitMergeIcon, GitPullRequestIcon, GitPullRequestClosedIcon, GitPullRequestDraftIcon} from '@primer/octicons-react';
 
 // This never changes, so it can be cached here
 export const getUsername = onetime(pageDetect.utils.getUsername);
@@ -117,10 +116,3 @@ export async function isPermalink(): Promise<boolean> {
 export function isNotRefinedGitHubRepo(): boolean {
 	return !location.pathname.startsWith('/sindresorhus/refined-github/');
 }
-
-export const stateIcon = {
-	OPEN: GitPullRequestIcon,
-	CLOSED: GitPullRequestClosedIcon,
-	MERGED: GitMergeIcon,
-	DRAFT: GitPullRequestDraftIcon,
-};
